requirements.txt
streamlit>=1.39.0
pymupdf>=1.24.0
ebooklib>=0.18
google-genai>=0.3.0
pydub>=0.25.1
kokoro-onnx>=0.5.0
soundfile>=0.12.1
python-dotenv>=1.0.1
